// src/app/core/elements/field.interface.ts
export interface Validator {
  name: string;
  validator: any;
  message: string;
}
export interface FieldConfig {
  label?: string;
  name?: string;
  inputType?: string;
  options?: { id: number, description: string }[];
  collections?: any;
  type: string;
  value?: any;
  validations?: Validator[];
}